Construct argument changed to reference
I was getting an Undefined Variable error each time adLDAP was loaded, appears to be as the $options array is only created in the constructor declaration, and isn't being returned beyond the function scope. Not sure if $options should be declared at the top of the class to keep everything tidy (would appreciate clarification) but PHP seems quite happy to muddle through with &$options as is.